+++ src/pif/checkers/ident/pif_ip_checker.py
@@ -1,18 +1,12 @@
+from __future__ import print_function
+
 __title__ = 'pif.checkers.ident.pif_ip_checker'
 __author__ = 'Artur Barseghyan'
 __copyright__ = 'Copyright (c) 2013 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('IdentmeIPChecker', 'V4IdentmeIPChecker', 'V6IdentmeIPChecker')
 
-from six import PY3
-
-try:
-    from six.moves.urllib.request import urlopen
-except ImportError as e:
-    if PY3:
-        from urllib.request import urlopen
-    else:
-        from urllib import urlopen
+from requests import get
 
 from pif.base import BasePublicIPChecker, registry
 
@@ -29,13 +23,18 @@ class IdentmeIPChecker(BasePublicIPChecker):
         :return str:
         """
         try:
-            data = str(urlopen('http://ident.me').read())
+            data = get('http://ident.me').text
             return data
         except Exception as e:
-            pass
+            if self.verbose:
+                print(e)
+
 
 registry.register(IdentmeIPChecker)
 
+# ********************************************************
+# ********************************************************
+# ********************************************************
 
 class V4IdentmeIPChecker(BasePublicIPChecker):
     """
@@ -50,13 +49,18 @@ class V4IdentmeIPChecker(BasePublicIPChecker):
         :return str:
         """
         try:
-            data = str(urlopen('http://v4.ident.me').read())
+            data = get('http://v4.ident.me').text
             return data
         except Exception as e:
-            pass
+            if self.verbose:
+                print(e)
+
 
 registry.register(V4IdentmeIPChecker)
 
+# ********************************************************
+# ********************************************************
+# ********************************************************
 
 class V6IdentmeIPChecker(BasePublicIPChecker):
     """
@@ -71,9 +75,11 @@ class V6IdentmeIPChecker(BasePublicIPChecker):
         :return str:
         """
         try:
-            data = str(urlopen('http://v6.ident.me').read())
+            data = get('http://v6.ident.me').text
             return data
         except Exception as e:
-            pass
+            if self.verbose:
+                print(e)
+
 
 registry.register(V6IdentmeIPChecker)